feature (core): move appInitService to core
