+++ projects/dsp-ui/src/lib/core/index.ts
@@ -1,2 +1,3 @@
 export * from './core.module';
 export * from './session.service';
+export * from './app-init.service';
